// gacha-expo/gacha-expo-frontend/lib/domain/shipping/types/shipping.type.ts
import { LineItem } from "@gacha-expo-app/domain/oripa/types/line-item";
import { UserAddress } from "@gacha-expo-app/domain/types";

export interface Shipping {
  id: string;
  userId?: string;
  addressId?: string;
  user: { email: string };
  address: UserAddress;
  addressInfo: string;
  lineItems: LineItem[];
  trackingNumber: string;
  status: string;
  createdAt : string;
  updatedAt: string;
  shippedAt: string;
}